--- examples/react/contexts/WalletSelectorContext.tsx
@@ -1,283 +1,0 @@
-import { setupCoin98Wallet } from "@near-wallet-selector/coin98-wallet";
-import type {
-  AccountState,
-  InjectedWalletBehaviour,
-  WalletSelector,
-} from "@near-wallet-selector/core";
-import { setupWalletSelector } from "@near-wallet-selector/core";
-import { setupHereWallet } from "@near-wallet-selector/here-wallet";
-import { setupMathWallet } from "@near-wallet-selector/math-wallet";
-import { setupMeteorWallet } from "@near-wallet-selector/meteor-wallet";
-import { setupNarwallets } from "@near-wallet-selector/narwallets";
-import type { WalletSelectorModal } from "@near-wallet-selector/modal-ui";
-import { setupModal } from "@near-wallet-selector/modal-ui";
-import { setupNearFi } from "@near-wallet-selector/nearfi";
-import { setupNightly } from "@near-wallet-selector/nightly";
-import { setupSender } from "@near-wallet-selector/sender";
-import { setupBitgetWallet } from "@near-wallet-selector/bitget-wallet";
-import { setupWalletConnect } from "@near-wallet-selector/wallet-connect";
-import { setupWelldoneWallet } from "@near-wallet-selector/welldone-wallet";
-import { setupNearSnap } from "@near-wallet-selector/near-snap";
-import { setupNeth } from "@near-wallet-selector/neth";
-import { setupMyNearWallet } from "@near-wallet-selector/my-near-wallet";
-import { setupLedger } from "@near-wallet-selector/ledger";
-import { setupXDEFI } from "@near-wallet-selector/xdefi";
-import { setupRamperWallet } from "@near-wallet-selector/ramper-wallet";
-import { setupNearMobileWallet } from "@near-wallet-selector/near-mobile-wallet";
-import { setupMintbaseWallet } from "@near-wallet-selector/mintbase-wallet";
-import { setupBitteWallet } from "@near-wallet-selector/bitte-wallet";
-import { setupOKXWallet } from "@near-wallet-selector/okx-wallet";
-import { setupEthereumWallets } from "@near-wallet-selector/ethereum-wallets";
-
-import type { ReactNode } from "react";
-import React, {
-  useCallback,
-  useContext,
-  useEffect,
-  useState,
-  useMemo,
-} from "react";
-import { distinctUntilChanged, map } from "rxjs";
-import { WagmiAdapter } from "@reown/appkit-adapter-wagmi";
-import { createAppKit } from "@reown/appkit/react";
-import { defineChain } from "@reown/appkit/networks";
-import type { CreateConnectorFn, GetAccountReturnType } from "@wagmi/core";
-import { reconnect, watchAccount } from "@wagmi/core";
-import { injected, walletConnect } from "@wagmi/connectors";
-
-import { Loading } from "../components/Loading";
-import { CONTRACT_ID } from "../constants";
-
-declare global {
-  interface Window {
-    selector: WalletSelector;
-    modal: WalletSelectorModal;
-  }
-}
-
-interface WalletSelectorContextValue {
-  selector: WalletSelector;
-  modal: WalletSelectorModal;
-  accounts: Array<AccountState>;
-  accountId: string | null;
-}
-
-const WalletSelectorContext =
-  React.createContext<WalletSelectorContextValue | null>(null);
-
-// Get a project ID at https://cloud.reown.com
-const projectId = "30147604c5f01d0bc4482ab0665b5697";
-
-const near = defineChain({
-  id: 398,
-  caipNetworkId: "eip155:398",
-  chainNamespace: "eip155",
-  name: "NEAR Protocol Testnet",
-  nativeCurrency: {
-    decimals: 18,
-    name: "NEAR",
-    symbol: "NEAR",
-  },
-  rpcUrls: {
-    default: { http: ["https://eth-rpc.testnet.near.org"] },
-    public: { http: ["https://eth-rpc.testnet.near.org"] },
-  },
-  blockExplorers: {
-    default: {
-      name: "NEAR Explorer",
-      url: "https://eth-explorer-testnet.near.org",
-    },
-  },
-  testnet: true,
-});
-
-const connectors: Array<CreateConnectorFn> = [
-  walletConnect({
-    projectId,
-    metadata: {
-      name: "NEAR Guest Book",
-      description: "A guest book with comments stored on the NEAR blockchain",
-      url: "https://near.github.io/wallet-selector",
-      icons: ["https://near.github.io/wallet-selector/favicon.ico"],
-    },
-    showQrModal: false, // showQrModal must be false
-  }),
-  injected({ shimDisconnect: true }),
-];
-
-const wagmiAdapter = new WagmiAdapter({
-  projectId,
-  connectors,
-  networks: [near],
-});
-reconnect(wagmiAdapter.wagmiConfig);
-
-const web3Modal = createAppKit({
-  adapters: [wagmiAdapter],
-  projectId,
-  networks: [near],
-  defaultNetwork: near,
-  enableWalletConnect: true,
-  features: {
-    analytics: true,
-    swaps: false,
-    onramp: false,
-    email: false, // Smart accounts (Safe contract) not available on NEAR Protocol, only EOA.
-    socials: false, // Smart accounts (Safe contract) not available on NEAR Protocol, only EOA.
-  },
-  coinbasePreference: "eoaOnly", // Smart accounts (Safe contract) not available on NEAR Protocol, only EOA.
-  allWallets: "SHOW",
-});
-
-export const WalletSelectorContextProvider: React.FC<{
-  children: ReactNode;
-}> = ({ children }) => {
-  const [selector, setSelector] = useState<WalletSelector | null>(null);
-  const [modal, setModal] = useState<WalletSelectorModal | null>(null);
-  const [accounts, setAccounts] = useState<Array<AccountState>>([]);
-  const [loading, setLoading] = useState<boolean>(true);
-
-  // Log in with Ethereum flow: auto connect to ethereum-wallets without showing the NEAR modal.
-  useEffect(() => {
-    if (!selector) {
-      return;
-    }
-    // Watch the connected Ethereum account and connect to the `ethereum-wallets` module automatically.
-    watchAccount(wagmiAdapter.wagmiConfig, {
-      onChange: (data: GetAccountReturnType) => {
-        if (!data.address || selector.store.getState().selectedWalletId) {
-          return;
-        }
-        selector.wallet("ethereum-wallets").then((wallet) =>
-          (wallet as InjectedWalletBehaviour).signIn({
-            contractId: CONTRACT_ID,
-          })
-        );
-      },
-    });
-  }, [selector]);
-
-  const init = useCallback(async () => {
-    const _selector = await setupWalletSelector({
-      network: "testnet",
-      debug: true,
-      modules: [
-        setupMyNearWallet(),
-        setupLedger(),
-        setupSender(),
-        setupBitgetWallet(),
-        setupMathWallet(),
-        setupNightly(),
-        setupMeteorWallet(),
-        setupNearSnap(),
-        setupOKXWallet(),
-        setupNarwallets(),
-        setupWelldoneWallet(),
-        setupHereWallet(),
-        setupCoin98Wallet(),
-        setupNearFi(),
-        setupRamperWallet(),
-        setupNeth({
-          gas: "300000000000000",
-          bundle: false,
-        }),
-        setupXDEFI(),
-        setupWalletConnect({
-          projectId: "c4f79cc...",
-          metadata: {
-            name: "NEAR Wallet Selector",
-            description: "Example dApp used by NEAR Wallet Selector",
-            url: "https://github.com/near/wallet-selector",
-            icons: ["https://avatars.githubusercontent.com/u/37784886"],
-          },
-        }),
-        setupNearMobileWallet(),
-        setupMintbaseWallet({ contractId: CONTRACT_ID }),
-        setupBitteWallet({ contractId: CONTRACT_ID }),
-        setupEthereumWallets({
-          wagmiConfig: wagmiAdapter.wagmiConfig,
-          web3Modal,
-        }),
-      ],
-    });
-    const _modal = setupModal(_selector, {
-      contractId: CONTRACT_ID,
-    });
-    const state = _selector.store.getState();
-    setAccounts(state.accounts);
-
-    // this is added for debugging purpose only
-    // for more information (https://github.com/near/wallet-selector/pull/764#issuecomment-1498073367)
-    window.selector = _selector;
-    window.modal = _modal;
-
-    setSelector(_selector);
-    setModal(_modal);
-    setLoading(false);
-  }, []);
-
-  useEffect(() => {
-    init().catch((err) => {
-      console.error(err);
-      alert("Failed to initialise wallet selector");
-    });
-  }, [init]);
-
-  useEffect(() => {
-    if (!selector) {
-      return;
-    }
-
-    const subscription = selector.store.observable
-      .pipe(
-        map((state) => state.accounts),
-        distinctUntilChanged()
-      )
-      .subscribe((nextAccounts) => {
-        console.log("Accounts Update", nextAccounts);
-
-        setAccounts(nextAccounts);
-      });
-
-    const onHideSubscription = modal!.on("onHide", ({ hideReason }) => {
-      console.log(`The reason for hiding the modal ${hideReason}`);
-    });
-
-    return () => {
-      subscription.unsubscribe();
-      onHideSubscription.remove();
-    };
-  }, [selector, modal]);
-
-  const walletSelectorContextValue = useMemo<WalletSelectorContextValue>(
-    () => ({
-      selector: selector!,
-      modal: modal!,
-      accounts,
-      accountId: accounts.find((account) => account.active)?.accountId || null,
-    }),
-    [selector, modal, accounts]
-  );
-
-  if (loading) {
-    return <Loading />;
-  }
-
-  return (
-    <WalletSelectorContext.Provider value={walletSelectorContextValue}>
-      {children}
-    </WalletSelectorContext.Provider>
-  );
-};
-
-export function useWalletSelector() {
-  const context = useContext(WalletSelectorContext);
-
-  if (!context) {
-    throw new Error(
-      "useWalletSelector must be used within a WalletSelectorContextProvider"
-    );
-  }
-
-  return context;
-}
