+++ examples/react/wallets/web3modal.ts
@@ -1,14 +1,17 @@
-import { createWeb3Modal } from "@web3modal/wagmi";
-
-import { reconnect, http, createConfig, type Config } from "@wagmi/core";
-import { type Chain } from "@wagmi/core/chains";
+import { WagmiAdapter } from "@reown/appkit-adapter-wagmi";
+import { createAppKit } from "@reown/appkit/react";
+import { defineChain } from "@reown/appkit/networks";
+import type { CreateConnectorFn } from "@wagmi/core";
+import { reconnect } from "@wagmi/core";
 import { injected, walletConnect } from "@wagmi/connectors";
 
-// Get a project ID at https://cloud.walletconnect.com
+// Get a project ID at https://cloud.reown.com
 const projectId = "30147604c5f01d0bc4482ab0665b5697";
 
-const near: Chain = {
+const near = defineChain({
   id: 398,
+  caipNetworkId: "eip155:398",
+  chainNamespace: "eip155",
   name: "NEAR Protocol Testnet",
   nativeCurrency: {
     decimals: 18,
@@ -26,32 +29,42 @@ const near: Chain = {
     },
   },
   testnet: true,
-};
-
-export const wagmiConfig: Config = createConfig({
-  chains: [near],
-  transports: {
-    [near.id]: http(),
-  },
-  connectors: [
-    walletConnect({
-      projectId,
-      metadata: {
-        name: "NEAR Guest Book",
-        description: "A guest book with comments stored on the NEAR blockchain",
-        url: "https://near.github.io/wallet-selector",
-        icons: ["https://near.github.io/wallet-selector/favicon.ico"],
-      },
-      showQrModal: false,
-    }),
-    injected({ shimDisconnect: true }),
-  ],
 });
-reconnect(wagmiConfig);
 
-export const web3Modal = createWeb3Modal({
-  wagmiConfig: wagmiConfig,
+export const connectors: Array<CreateConnectorFn> = [
+  walletConnect({
+    projectId,
+    metadata: {
+      name: "NEAR Guest Book",
+      description: "A guest book with comments stored on the NEAR blockchain",
+      url: "https://near.github.io/wallet-selector",
+      icons: ["https://near.github.io/wallet-selector/favicon.ico"],
+    },
+    showQrModal: false, // showQrModal must be false
+  }),
+  injected({ shimDisconnect: true }),
+];
+
+export const wagmiAdapter = new WagmiAdapter({
   projectId,
-  enableOnramp: false,
+  connectors,
+  networks: [near],
+});
+reconnect(wagmiAdapter.wagmiConfig);
+
+export const web3Modal = createAppKit({
+  adapters: [wagmiAdapter],
+  projectId,
+  networks: [near],
+  defaultNetwork: near,
+  enableWalletConnect: true,
+  features: {
+    analytics: true,
+    swaps: false,
+    onramp: false,
+    email: false, // Smart accounts (Safe contract) not available on NEAR Protocol, only EOA.
+    socials: false, // Smart accounts (Safe contract) not available on NEAR Protocol, only EOA.
+  },
+  coinbasePreference: "eoaOnly", // Smart accounts (Safe contract) not available on NEAR Protocol, only EOA.
   allWallets: "SHOW",
 });
